Refactor layered plan and view.
