+++ include/mneme/view.hpp
@@ -7,6 +7,7 @@
 #include <vector>
 
 #include "iterator.hpp"
+#include "util.hpp"
 
 namespace mneme {
 
@@ -90,6 +91,74 @@ private:
     Storage& container;
     offset_type offset;
 };
+
+template <typename MaybeStride = StaticNothing, typename MaybePlan = StaticNothing,
+          typename MaybeStorage = StaticNothing>
+class LayeredViewFactory {
+
+public:
+    LayeredViewFactory() = default;
+
+    LayeredViewFactory(MaybePlan maybePlan, MaybeStorage maybeStorage)
+        : maybePlan(maybePlan), maybeStorage(maybeStorage) {}
+
+    template <std::size_t Stride>[[nodiscard]] auto withStride() const {
+        return LayeredViewFactory<std::integral_constant<size_t, Stride>, MaybePlan, MaybeStorage>(
+            maybePlan, maybeStorage);
+    }
+
+    [[nodiscard]] auto withDynamicStride() const { return withStride<dynamic_extent>(); }
+
+    template <typename LayeredPlanT> auto withPlan(LayeredPlanT& plan) const {
+        const auto somePlan = StaticSome<LayeredPlanT>(plan);
+        return LayeredViewFactory<MaybeStride, StaticSome<LayeredPlanT>, MaybeStorage>(
+            somePlan, maybeStorage);
+    }
+
+    template <typename StorageT> constexpr auto withStorage(StorageT& storage) const {
+        const auto someStorage = StaticSome<StorageT*>(&storage);
+        return LayeredViewFactory<MaybeStride, MaybePlan, StaticSome<StorageT*>>(maybePlan,
+                                                                                 someStorage);
+    }
+
+    template <
+        typename Layer, typename MaybeStride_ = MaybeStride, typename MaybePlan_ = MaybePlan,
+        typename MaybeStorage_ = MaybeStorage,
+        typename std::enable_if<!std::is_same<MaybeStride_, StaticNothing>::value, int>::type = 0,
+        typename std::enable_if<!std::is_same<MaybePlan_, StaticNothing>::value, int>::type = 0,
+        typename std::enable_if<!std::is_same<MaybeStorage_, StaticNothing>::value, int>::type = 0>
+    [[nodiscard]] auto createStridedView() const {
+        auto layout = maybePlan.value.getLayout();
+        const auto& layer = maybePlan.value.template getLayer<Layer>();
+        size_t from = layer.offset;
+        size_t to = from + layer.numElements;
+        return StridedView<std::remove_pointer_t<typename MaybeStorage_::type>, MaybeStride::value>(
+            layout, *(maybeStorage.value), from, to);
+    }
+
+    template <
+        typename Layer, typename MaybeStride_ = MaybeStride, typename MaybePlan_ = MaybePlan,
+        typename MaybeStorage_ = MaybeStorage,
+        typename std::enable_if<std::is_same<MaybeStride_, StaticNothing>::value, int>::type = 0,
+        typename std::enable_if<!std::is_same<MaybePlan_, StaticNothing>::value, int>::type = 0,
+        typename std::enable_if<!std::is_same<MaybeStorage_, StaticNothing>::value, int>::type = 0>
+    constexpr auto createDenseView() {
+        auto layout = maybePlan.value.getLayout();
+        const auto& layer = maybePlan.value.template getLayer<Layer>();
+        size_t from = layer.offset;
+        size_t to = from + layer.numElements;
+        return DenseView<std::remove_pointer_t<typename MaybeStorage_::type>>(
+            layout, *(maybeStorage.value), from, to);
+    }
+
+private:
+    MaybePlan maybePlan;
+    MaybeStorage maybeStorage;
+};
+
+constexpr auto createView() {
+    return LayeredViewFactory<StaticNothing, StaticNothing, StaticNothing>();
+}
 
 } // namespace mneme
